+++ examples/flutter_gpu_shim_smoke/lib/triangle_tab.dart
@@ -122,14 +122,7 @@ class _TriangleTabState extends State<TriangleTab> {
         gpu.IndexType.int16,
         6,
       );
-      pass.setViewport(
-        gpu.Viewport(
-          x: 0,
-          y: 0,
-          width: _size.toDouble(),
-          height: _size.toDouble(),
-        ),
-      );
+      pass.setViewport(gpu.Viewport(x: 0, y: 0, width: _size, height: _size));
       pass.draw();
       pass.clearBindings();
       cmd.submit();
